requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.1.1
openapi-codec==1.3.2
PyMySQL==1.0.2
pytz==2022.7.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.3
